+++ scripts/CGR_functions.py
