+++ scripts/cor_dist_feature.R
@@ -6,18 +6,25 @@
 
 ### Correlation analysis of a a feature in function of distance to median region
 # Will input the arguments:
-# 1. path to the distance matrix file
-# 2. path to the feature file
-# 3. feature as displayed in the title and y axis
-# 4. distance matrix information (type of genomic signature/type of distance/window size)
-# 5. path to the output image (with its name in the path)
+# 1. path to the output image (with its name in the path)
+# 2. NOT USED IN THIS SCRIPT
+# 3. path to the distance matrix
+# 4. path to the feature file
+# 5. feature type
+# 6. window size
+# 7. Genomic signature type
+# 8. IF USING FCGR : k-mer size
 args <- commandArgs(trailingOnly=TRUE)
 
-feature_type = basename(args[3])
-window_size = basename(args[4])
-output = args[5]
+output = args[1]
+feature_type = basename(args[5])
+window_size = basename(args[6])
+gs = basename(args[7])
+if (gs == 'FCGRs'){
+  kmer = basename(args[8])
+} 
 
-distance_matrix <- read.table(args[1], sep = "\t", header = FALSE)
+distance_matrix <- read.table(args[3], sep = "\t", header = FALSE)
 # Will remove any empty column
 distance_matrix <- distance_matrix[, colSums(is.na(distance_matrix)) == 0]
 
@@ -30,16 +37,24 @@ min_column = which.min(colSums(distance_matrix))
 median_region = distance_matrix[-min_column, min_column]
 
 
-feature <- read.table(args[2], sep = "\t", header = FALSE)
+feature <- read.table(args[4], sep = "\t", header = FALSE)
 # We take out the median region's feature
 feature <- feature[-min_column,]
 
 # Quick correlation test and parsing the resulting p-value
 pv <- round(cor.test(median_region, feature[,2])$p.value, 6)
 
-plot_title <- paste('% of', feature_type, 'in function of distance to the median region \n',
-                    paste('(FCGRs, ', window_size, ' bp)', '\n', sep = ''),
-                    'Correlation test p-value :', pv)
+if (gs == 'FCGRs'){
+  plot_title <- paste('% of', feature_type, 'in function of distance to the median region \n',
+                      paste('(FCGRs, k= ', kmer, ', ', window_size, ' bp windows)', '\n', sep = ''),
+                      'Correlation test p-value :', pv)
+
+} else {
+  plot_title <- paste('% of', feature_type, 'in function of distance to the median region \n',
+                      paste('(DFTs, ', window_size, ' bp windows)', '\n', sep = ''),
+                      'Correlation test p-value :', pv)
+}
+
 
 png(output, width=700, height=500, units="px")
 plot(median_region, feature[,2], main=plot_title, xlab='Distance to median region',
