+++ scripts/windowed_CGR.py
